extract dataset ability into separate file
