+++ src/casl/casl.module.ts
@@ -2,9 +2,11 @@ import { Module } from "@nestjs/common";
 import { ConfigModule } from "@nestjs/config";
 import { CaslAbilityFactory } from "./casl-ability.factory";
 import { JobConfigModule } from "src/config/job-config/jobconfig.module";
+import { DatasetAbility } from "./abilities/datasets.ability";
+
 @Module({
   imports: [JobConfigModule, ConfigModule],
-  providers: [CaslAbilityFactory],
+  providers: [CaslAbilityFactory, DatasetAbility],
   exports: [CaslAbilityFactory],
 })
 export class CaslModule {}
